pyproject.toml
[project]
name = "cityscope-backend"
version = "0.1.0"
description = "Backend API for CityScope project (FastAPI + JWT + Postgres)"
authors = [
  { name = "Ricardo Wurzmann", email = "cadu@kovrr.com" }
]
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
  "fastapi>=0.115",
  "uvicorn[standard]>=0.30",
  "python-dotenv>=1.0",
  "pydantic>=2.8",
  "sqlalchemy>=2.0",
  "psycopg[binary]>=3.2",
  "alembic>=1.13",
  "passlib[bcrypt]>=1.7",
  "bcrypt>=4.0.0",
  "pyjwt[crypto]>=2.9",
  "httpx>=0.27",
  "email-validator>=2.1",
]

[project.optional-dependencies]
dev = [
  "black>=24.0",
  "ruff>=0.6",
  "pytest>=8.0",
]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.ruff]
line-length = 100
target-version = "py311"
select = ["E", "F", "I"]
ignore = ["E501"]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"
